Update DiskSpace.java
Made `someFile` which is of type RandomAccessFile that we can use throughout the process. Also, changed up the `seek` and `write` in `void write()` to fit the descriptor of what we are writing. Cleaned up the excess comments for turn in. And made a comment on `area*1024`. I am pretty sure that this is to large, but I would like to here other's opinions.